clean up code

+++ app/src/main/java/com/BlizzardArmory/warcraft/encounters/Encounters.kt
@@ -6,5 +6,5 @@ data class Encounters(
 
         @SerializedName("encounter") val encounter: Encounter,
         @SerializedName("completed_count") val completed_count: Int,
-        @SerializedName("last_kill_timestamp") val last_kill_timestamp: Int
+        @SerializedName("last_kill_timestamp") val last_kill_timestamp: Long
 )
