+++ app/src/main/java/com/BlizzardArmory/warcraft/encounters/Instance.kt
@@ -6,5 +6,5 @@ data class Instance(
 
         @SerializedName("key") val key: Key,
         @SerializedName("name") val name: String,
-        @SerializedName("id") val id: Int
+        @SerializedName("id") val id: Long
 )
